[module 2] revise PR: changing folder test life cycle
