+++ modules/test/test.typedefs.js
@@ -55,6 +55,10 @@ const testTypeDefs = gql`
     max_point: Float!
   }
 
+  input AssignCorrectorInput {
+    user_id: ID!
+  }
+
   extend type Query {
     GetAllTests(published_status: PublishedStatus): [Test]
     GetOneTest(_id: ID!): Test
@@ -64,6 +68,7 @@ const testTypeDefs = gql`
     CreateTest(test_input: TestInput): Test!
     UpdateTest(_id: ID!, test_input: TestInput): Test
     PublishTest(task_input: PublishTestInput!): ID
+    AssignCorrector(_id: ID!, task_input: AssignCorrectorInput!): ID
     DeleteTest(_id: ID!): ID
   }
 `;
